Add bar as visualisation choice
Adds a horizontal bar to the token showing the value as portion of the
max value for a counter. A value higher than max will always show a
full bar and zero or lower will show an empty bar, as will a counter
with the max value zero or lower.

Changes how exports are defined. Now their done for each exported
component rather than at the head of the file. The visualistaion
module is now namespaced and function names have been changed
accordingly.