Add field for modifying the value using plus or minus expressions
It's assumed that the value in the "modify" field starts with "+" or "-"
followed by a number. If this is not fulfilled nothing will happen. On
the "change" event the "modify" value will be applied to the "value" field
and then updated as if a change was made to that field.

The value of the modify field is not stored when the UI updates.